refactor(gpio): delete gpio_helpers.py, adapt tests for gpio_client
Remove the now-unused gpio_helpers.py module. All production code has
been migrated to gpio_client in previous commits.

Rename test_gpio_helpers.py to test_gpio_client_compat.py, retaining
18 tests for pure-logic functions (get_relay_level, _is_active_low,
_pin_to_relay_name, read_gpio_state file fallback). Remove 25 tests
that tested RPi.GPIO-specific behavior (setup_relay, relay_on/off via
GPIO.output, write_gpio_state to file, GPIO_AVAILABLE flag) — these
are already covered by mock-socket tests in test_gpio_client.py.
